fix type cast bug

+++ core/main_loop.hpp
@@ -46,7 +46,7 @@ inline void Node::MainLoop() {
                 if constexpr (std::is_same_v<T, AppendEntriesReqPayload>) {
                     #ifdef DEBUG
                     std::cout << "found AE RPC from node " << payload.leader_id << "\n";
-                    std::cout << "state_ = " << state_ << "\n";
+                    std::cout << "state_ = " << static_cast<int>(state_) << "\n";
                     std::cout << "payload term = " << payload.term << "\n";
                     std::cout << "current term = " << current_term_ << "\n";
                     std::cout << "entries_len = " << payload.entries_len << "\n";
